socialenrollr/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
Werkzeug==3.0.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
